Errors Fixed, Logic Issues Resolved, Maven Configuration Problems Resolved
