+++ common/src/main/java/mechanism/SearchMechanism.java
@@ -1,0 +1,13 @@
+package mechanism;
+
+/**
+ Represents a search mechanism for finding people based on certain criteria.
+ */
+public interface SearchMechanism {
+
+    /**
+     * Finds a list of people based on the defined search criteria.
+     */
+    void findPeople();
+
+}
