+++ common/src/main/java/model/Person.java
@@ -1,0 +1,12 @@
+package model;
+
+/**
+ * model.Person class designed
+ */
+public record Person(String firstName, String lastName, String email) {
+
+    @Override
+    public String toString() {
+        return firstName + " " + lastName + " " + email;
+    }
+}
