--- src/main/java/search/SearchMechanism.java
@@ -1,13 +1,0 @@
-package search;
-
-/**
- Represents a search mechanism for finding people based on certain criteria.
- */
-public interface SearchMechanism {
-
-    /**
-     * Finds a list of people based on the defined search criteria.
-     */
-    void findPeople();
-
-}
